Add rule, validator tests. Test exact file contents generated.
