+++ tests/Actions/GenerateActionTest.php
@@ -8,15 +8,12 @@ use PerfectOblivion\ActionServiceResponder\Tests\TestCase;
 
 class GenerateActionTest extends TestCase
 {
-    /**
-     * @test
-     *
-     * The asr:action command will generate a new action in the
-     * configured namespace from the provided Action stub.
-     */
+    /** @test */
     function the_action_command_creates_an_action_from_stub()
     {
-        $action = app_path($this->paths['action'].'/'.'User/StoreUser.php');
+        $testStub = str_replace("\r\n", "\n", File::get(__DIR__.'/../stubs/action.stub'));
+
+        $action = app_path($this->paths['action'].'/'.'Sample/ActionSample.php');
 
         if (File::exists($action)) {
             unlink($action);
@@ -24,8 +21,9 @@ class GenerateActionTest extends TestCase
 
         $this->assertFalse(File::exists($action));
 
-        Artisan::call('asr:action User\\\StoreUser');
+        Artisan::call('asr:action Sample\\\ActionSample');
 
         $this->assertTrue(File::exists($action));
+        $this->assertEquals($testStub, str_replace("\r\n", "\n", File::get($action)));
     }
 }
