+++ tests/Foundation/Actions/SimpleResponderAction.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace PerfectOblivion\ActionServiceResponder\Tests\Foundation\Actions;
+
+use PerfectOblivion\ActionServiceResponder\Actions\Action;
+use PerfectOblivion\ActionServiceResponder\Tests\Foundation\Responders\SimpleResponder;
+
+class SimpleResponderAction extends Action
+{
+    /** @var \PerfectOblivion\ActionServiceResponder\Tests\Foundation\Responders\SimpleResponder */
+    private $responder;
+
+    /**
+     * Construct a new SimpleResponderAction
+     *
+     * @param  \PerfectOblivion\ActionServiceResponder\Tests\Foundation\Responders\SimpleResponder  $responder
+     */
+    public function __construct(SimpleResponder $responder)
+    {
+        $this->responder = $responder;
+    }
+
+    /**
+     * Execute the action.
+     *
+     *  @param  \Illuminate\Http\Request  $request
+     */
+    public function __invoke(): string
+    {
+        return $this->responder->respond();
+    }
+}
