+++ tests/Foundation/Requests/CustomRequest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace PerfectOblivion\ActionServiceResponder\Tests\Foundation\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class CustomRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            //
+        ];
+    }
+}
