+++ tests/Foundation/Responders/SimpleResponderWithPayload.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace PerfectOblivion\ActionServiceResponder\Tests\Foundation\Responders;
+
+use PerfectOblivion\ActionServiceResponder\Responders\Responder;
+
+class SimpleResponderWithPayload extends Responder
+{
+    /**
+     * Send a response.
+     *
+     * @return mixed
+     */
+    public function respond()
+    {
+        return $this->payload;
+    }
+}
