+++ tests/Foundation/Services/TestServiceValidator.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace PerfectOblivion\ActionServiceResponder\Tests\Foundation\Services;
+
+use PerfectOblivion\ActionServiceResponder\Validation\ValidationService\ValidationService;
+
+class TestServiceValidator extends ValidationService
+{
+    /**
+     * Get the validation rules that apply to the data.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'name' => ['required', 'string'],
+        ];
+    }
+}
