+++ tests/Responders/GenerateResponderTest.php
@@ -4,6 +4,7 @@ namespace PerfectOblivion\ActionServiceResponder\Tests\Responders;
 
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 use PerfectOblivion\ActionServiceResponder\Tests\TestCase;
 
 class GenerateResponderTest extends TestCase
@@ -11,7 +12,9 @@ class GenerateResponderTest extends TestCase
     /** @test */
     function the_responder_command_creates_an_responder()
     {
-        $responder = app_path($this->paths['responder'].'/'.'User/StoreUserResponder.php');
+        $testStub = str_replace("\r\n", "\n", File::get(__DIR__.'/../stubs/responder.stub'));
+
+        $responder = app_path($this->paths['responder'].'/'.'Sample/SampleResponder.php');
 
         if (File::exists($responder)) {
             unlink($responder);
@@ -19,8 +22,9 @@ class GenerateResponderTest extends TestCase
 
         $this->assertFalse(File::exists($responder));
 
-        Artisan::call('asr:responder User\\\StoreUserResponder');
+        Artisan::call('asr:responder Sample\\\SampleResponder');
 
         $this->assertTrue(File::exists($responder));
+        $this->assertEquals($testStub, str_replace("\r\n", "\n", File::get($responder)));
     }
 }
