--- tests/Stubs/Actions/SimpleAction.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace PerfectOblivion\ActionServiceResponder\Tests\Stubs\Actions;
-
-use PerfectOblivion\ActionServiceResponder\Actions\Action;
-
-class SimpleAction extends Action
-{
-    /**
-     * Execute the action.
-     *
-     *  @param  \Illuminate\Http\Request  $request
-     */
-    public function __invoke()
-    {
-        return 'hello';
-    }
-}
