--- tests/Stubs/Services/TestServiceObjectPayload.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace PerfectOblivion\ActionServiceResponder\Tests\Stubs\Services;
-
-use PerfectOblivion\ActionServiceResponder\Services\Service;
-use PerfectOblivion\ActionServiceResponder\Services\Traits\SelfCallingService;
-
-class TestServiceObjectPayload extends Service
-{
-    use SelfCallingService;
-
-    /**
-     * Handle the call to the service.
-     *
-     * @param  array  $parameters
-     *
-     * @return mixed
-     */
-    public function run(array $parameters)
-    {
-        return $parameters;
-    }
-}
