--- tests/Stubs/Services/TestServiceValidator.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace PerfectOblivion\ActionServiceResponder\Tests\Stubs\Services;
-
-use PerfectOblivion\ActionServiceResponder\Validation\ValidationService\ValidationService;
-
-class TestServiceValidator extends ValidationService
-{
-    /**
-     * Get the validation rules that apply to the data.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'name' => ['required', 'string'],
-        ];
-    }
-}
