+++ tests/stubs/responder.stub
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Http\Responders\Sample;
+
+use PerfectOblivion\ActionServiceResponder\Responders\Responder;
+
+class SampleResponder extends Responder
+{
+    /**
+     * Send a response.
+     *
+     * @return mixed
+     */
+    public function respond()
+    {
+        //
+    }
+}
