+++ tests/stubs/rule.stub
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Rules;
+
+use PerfectOblivion\ActionServiceResponder\Validation\CustomRule;
+
+class SampleRule extends CustomRule
+{
+    /**
+     * Determine if the validation rule passes.
+     *
+     * @param  string  $attribute
+     * @param  mixed  $value
+     * @return bool
+     */
+    public function passes($attribute, $value)
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation error message.
+     *
+     * @return string
+     */
+    public function message()
+    {
+        return 'The validation error message.';
+    }
+}
